# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CRM monitoring API for arbitrage trading system.",
        "title": "Arbitrage VM CRM API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/system/exchanges": {
            "get": {
                "description": "List supported exchanges and latest market-data timestamps",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "System exchanges",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SystemExchangeList"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/system/status": {
            "get": {
                "description": "Get API, database, TimescaleDB, market-data, allocation, and exchange status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "System status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SystemStatus"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.SystemStatus"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.HealthComponent": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.SystemAllocationStatus": {
            "type": "object",
            "properties": {
                "last_updated_at": {
                    "type": "string"
                },
                "running": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.SystemExchange": {
            "type": "object",
            "properties": {
                "credential_configured": {
                    "type": "boolean"
                },
                "demo": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "exchange": {
                    "type": "string"
                },
                "last_account_event_at": {
                    "type": "string"
                },
                "last_funding_at": {
                    "type": "string"
                },
                "last_market_data_at": {
                    "type": "string"
                },
                "last_market_quality_at": {
                    "type": "string"
                },
                "last_open_interest_at": {
                    "type": "string"
                },
                "last_wallet_snapshot_at": {
                    "type": "string"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "supported": {
                    "type": "boolean"
                }
            }
        },
        "response.SystemExchangeList": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.SystemExchange"
                    }
                }
            }
        },
        "response.SystemMarketDataStatus": {
            "type": "object",
            "properties": {
                "funding_last_at": {
                    "type": "string"
                },
                "market_quality_last_at": {
                    "type": "string"
                },
                "open_interest_last_at": {
                    "type": "string"
                }
            }
        },
        "response.SystemStatus": {
            "type": "object",
            "properties": {
                "allocations": {
                    "$ref": "#/definitions/response.SystemAllocationStatus"
                },
                "api_version": {
                    "type": "string"
                },
                "database": {
                    "$ref": "#/definitions/response.HealthComponent"
                },
                "exchanges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.SystemExchange"
                    }
                },
                "market_data": {
                    "$ref": "#/definitions/response.SystemMarketDataStatus"
                },
                "status": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "timescale": {
                    "$ref": "#/definitions/response.HealthComponent"
                }
            }
        }
    }
}